requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
transformers>=4.40.0
torch>=2.1.0
pillow>=10.1.0
reportlab>=4.0.7
ebooklib>=0.18
diffusers>=0.27.0
accelerate>=0.30.0
sentencepiece>=0.1.99
protobuf>=4.25.1
requests>=2.31.0
numpy>=1.24.3
huggingface-hub>=0.19.4

